pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "upskill_python_exercise"
version = "0.1"
requires-python = ">=3.7"
dependencies = [
    "sqlalchemy>=1.4",
    "python-dotenv>=0.19",
]

[tool.setuptools]
packages = ["core", "sql_alchemy", "sftp_api", "sites_api"]